+++ pkgs/artisanal_widgets/test/input/text_area_test.dart
@@ -1,8 +1,8 @@
 import 'package:artisanal/bubbles.dart' show TextAreaModel;
 import 'package:artisanal/terminal.dart' show KeyType;
+import 'package:artisanal/testing.dart';
 import 'package:artisanal/tui.dart' as tui;
-import 'package:artisanal_widgets/artisanal_widgets.dart';
-import 'package:artisanal_widgets/testing.dart';
+import 'package:artisanal/widgets.dart';
 import 'package:test/test.dart';
 
 void main() {