+++ pkgs/artisanal_widgets/test/input/text_field_multiline_test.dart
@@ -1,9 +1,9 @@
 import 'package:artisanal/terminal.dart' show KeyType;
 import 'package:artisanal/terminal.dart' as terminal show Key;
+import 'package:artisanal/testing.dart';
 import 'package:artisanal/tui.dart'
     show KeyMsg, MouseAction, MouseButton, MouseMsg;
-import 'package:artisanal_widgets/artisanal_widgets.dart';
-import 'package:artisanal_widgets/testing.dart';
+import 'package:artisanal/widgets.dart';
 import 'package:test/test.dart';
 
 void main() {